task completed https://up.htmlacademy.ru/ecmascript/10/tasks/4
